--- SharePointFramework/ProjectExtensions/src/extensions/projectSetup/IProjectSetupProperties.ts
@@ -1,31 +1,0 @@
-export interface IProjectSetupProperties {
-  /**
-   * Templates library
-   */
-  templatesLibrary: string
-
-  /**
-   * Exensions library
-   */
-  extensionsLibrary: string
-
-  /**
-   * Projects list
-   */
-  projectsList: string
-
-  /**
-   * List content config list
-   */
-  contentConfigList: string
-
-  /**
-   * Term set IDs
-   */
-  termSetIds: { [key: string]: string }
-
-  /**
-   * Tasks
-   */
-  tasks: string[]
-}
