--- SharePointFramework/ProjectExtensions/src/extensions/projectSetup/ProjectSetupValidation.ts
@@ -1,6 +1,0 @@
-export enum ProjectSetupValidation {
-  InvalidWebLanguage,
-  NoHubConnection,
-  InvalidCulture,
-  Ready
-}
