+++ SharePointFramework/ProjectExtensions/src/tasks/ApplyTemplate/index.ts
@@ -3,7 +3,7 @@ import { format } from 'office-ui-fabric-react/lib/Utilities'
 import * as strings from 'ProjectExtensionsStrings'
 import { Web, WebProvisioner } from 'sp-js-provisioning'
 import * as _ from 'underscore'
-import { IProjectSetupData } from '../../extensions/projectSetup/IProjectSetupData'
+import { IProjectSetupData } from '../../extensions/projectSetup/types'
 import { BaseTask, BaseTaskError, IBaseTaskParams } from '../@BaseTask'
 import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
 import { APPLY_TEMPLATE_STATUS_MAP } from './ApplyTemplateStatusMap'
